migrate narrative tools out of agent module
